Fixed all require entries and do_ship adjustment
Many recipes were using require and pointing to the actual location of
the inc file.  Change all non-GHC libs to just reference the file.  For
GHC libs those includes were all turned into classes so set them to
inherit the proper class.  Also tweaked how the do_ship task was added
in the image recipes to make sure it is as far at the end of a recipe
as possible.