Fix calling of native binaries
OpenXT relied on the PATH being set correctly to call native binaries. To
better align with being an OE layer we should instead specifically state
when we need the native version.  Maybe this can be put in a
selinuxnative class later like python.  For now, just add the proper
variables.