Work to support debugging
Some ASSNEEDED tweaks to get packages to compile as a debug build.  When
doing either a debug build or inhibiting the package strip, the resulting
images will be much larger.  The settings to tweak this have been left in
but commented out.  The sizes are not absolute and may vary depending on
the package versions used and whether doing a debug build or just
inhibiting stripping.

Down the road an arbitrarily large value for each could be selected and
set by detecting which variables were set.  In addition to these it will
be necessary to increase the partition sizes in the installer package. If
not done the installation of OpenXT will fail due to the lvm partition
being full.