libevent forced to use -lrt
While troublshooting I noticed dm-agent was not linking against librt as
in the original OpenXT.  This eventually led me to seeing that libevent
was what was not being linked.  The test for clock_gettime in libevent
seems to think libevent never uses that call even though I see it in the
code.  Until the cause is determined I have just manually set the -lrt
flag.